move into localize directory (#3470)
* move into localize directory

* rename LocalizeMixin -> _LocalizeMixinBase and LocalizeDynamicMixin -> LocalizeMixin

* handle synchronous importFunc
