+++ helpers/localize-core-element.js
@@ -1,6 +1,6 @@
-import { LocalizeDynamicMixin } from '../mixins/localize-dynamic-mixin.js';
+import { LocalizeMixin } from '../mixins/localize/localize-mixin.js';
 
-export const LocalizeCoreElement = superclass => class extends LocalizeDynamicMixin(superclass) {
+export const LocalizeCoreElement = superclass => class extends LocalizeMixin(superclass) {
 
 	static get localizeConfig() {
 		return {
