+++ mixins/localize/demo/localize-mixin.html
@@ -3,9 +3,9 @@
 	<head>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<meta charset="UTF-8">
-		<link rel="stylesheet" href="../../components/demo/styles.css" type="text/css">
+		<link rel="stylesheet" href="../../../components/demo/styles.css" type="text/css">
 		<script type="module">
-			import '../../components/demo/demo-page.js';
+			import '../../../components/demo/demo-page.js';
 			import './localize-mixin-greeting.js';
 			import './localize-mixin-mission.js';
 		</script>
